UPDATE - Move IntoDiagnosticArg implementations to diagnostic_impls file
